LayoutTests/media/modern-media-controls/audio/audio-controls-metrics-expected.txt

Testing the <audio> controls metrics.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bounds.width is 200
PASS bounds.height is 50

PASS successfullyParsed is true

TEST COMPLETE
